fix details about color config refactor
