+++ Ghidra/Framework/Docking/src/main/java/docking/widgets/EmptyBorderButton.java
@@ -24,12 +24,11 @@ import javax.swing.border.Border;
 import javax.swing.border.EmptyBorder;
 import javax.swing.event.ChangeEvent;
 import javax.swing.event.ChangeListener;
-import java.awt.*;
 import java.awt.event.MouseAdapter;
 import java.awt.event.MouseEvent;
 
 import static javax.swing.BorderFactory.createEmptyBorder;
-import static ghidra.util.config.ReadConfigProperties.findColor;
+import static ghidra.framework.ColorProperties.findColor;
 
 /**
  * Class that is a JButton that has an empty border and adds a mouse listener
