+++ Ghidra/RuntimeScripts/Common/support/Color.properties
@@ -168,7 +168,7 @@ Dark.Theme.White = #FF383838
 Background.Default = ${White}
 Background = ${White}
 Foreground = ${Black}
-Based.Background.Model = ${Dark.Theme.White}
+Listing.Background = ${Dark.Theme.White}
 
 # ::Selection | highlight
 Selection = #244570
@@ -253,8 +253,8 @@ Pcode.Blue = #B5C5F5
 Pcode.Red = #FB4747
 
 # ---- MnemonicFieldFactory ----
-Override = #DB51A4
-ProtoType.Bad = #CC3737
+# Mnemonic.Override = #DB51A4
+Mnemonic.BadPrototype = #CC3737
 
 # ==== Components ====
 # ---- Button ----
